* master: (758 commits)
  svn update
  update addon scripts
  update translation files
  ndof: restored support for older serial port 3D mice on Linux (R45951 broke it, oops)
  SVN maintenance.
  adding kernel_projection.h in CMakelists
  Cycles: add light falloff node, with quadratic/linear/constant falloff and a smoothing factor to reduce high values near the light. http://wiki.blender.org/index.php/Doc:2.6/Manual/Render/Cycles/Nodes/More#Light_Falloff
  Collada: patch #31331: added missing call to BKE_mesh_tessface_ensure() when 'apply Modifiers' was NOT selected
  style cleanup: avi
  fix [#31354] Segfault with "Extrude Individual Faces and Move"
  Collada: patch #31331: Implementation of 'Apply Modifiers'
  fix [#31347] solidify modifier > even thickness parameter weired
  style cleanup: outliner
  cycles-fisheye: fixed formula for dir to equisolid second time I need a scond commit to get the equisolid formula right, shame on me ;) the formula is: r = 2 x focallength x sin (theta / 2)
  direction to equisolid - final fix for [#31307] Cycles panoramic fisheye lens and vector pass problem
  Some misc fixes to UI messages...
  fix for solidify modifier angle calculation (was incorrectly using normalized angle function)
  GHOST/Cocoa: revert tablet detection commit for continuous grab, apparently it is already doing it and this change breaks something? Still it seems strange that it is doing tablet stuff in this part of the code then.
  GHOST/Cocoa: detect tablet event to disable continuous grab. Don't have tablet to test it though, can someone with Mac + tablet confirm if continuous grab gets automatically disabled when using the tablet?
  Fix most of #31307: cycles panorama camera not working correct with speed vectors and window texture coordinates. Only for Fisheye Equisolid it's still not working correct yet. Patch from Dalai with modifications.
  ...

Conflicts:
	source/blender/editors/space_outliner/outliner_tree.c

+++ source/blender/editors/transform/transform_input.c
@@ -199,17 +199,17 @@ static void InputCustomRatio(TransInfo *UNUSED(t), MouseInput *mi, const int mva
 			mdx = (mi->precision_mval[0] + (float)(mval[0] - mi->precision_mval[0]) / 10.0f) - data[2];
 			mdy = (mi->precision_mval[1] + (float)(mval[1] - mi->precision_mval[1]) / 10.0f) - data[3];
 
-			distance = (length != 0.0f)? (mdx*dx + mdy*dy) / length: 0.0f;
+			distance = (length != 0.0) ? (mdx * dx + mdy * dy) / length: 0.0;
 		}
 		else {
 			int mdx, mdy;
 			mdx = mval[0] - data[2];
 			mdy = mval[1] - data[3];
 
-			distance = (length != 0.0f)? (mdx*dx + mdy*dy) / length: 0.0f;
+			distance = (length != 0.0) ? (mdx * dx + mdy * dy) / length: 0.0;
 		}
 
-		output[0] = (float)((length != 0.0f)? distance / length: 0.0f);
+		output[0] = (length != 0.0) ? (double)(distance / length) : 0.0f;
 	}
 }
 
@@ -313,8 +313,7 @@ void initMouseInputMode(TransInfo *t, MouseInput *mi, MouseInputMode mode)
 	}
 #endif
 
-	switch(mode)
-	{
+	switch (mode) {
 	case INPUT_VECTOR:
 		mi->apply = InputVector;
 		t->helpline = HLP_NONE;
@@ -391,8 +390,7 @@ int handleMouseInput(TransInfo *t, MouseInput *mi, wmEvent *event)
 {
 	int redraw = TREDRAW_NOTHING;
 
-	switch (event->type)
-	{
+	switch (event->type) {
 	case LEFTSHIFTKEY:
 	case RIGHTSHIFTKEY:
 		if (event->val == KM_PRESS) {
